test/e2e/framework: remove skip.go and use e2eskipper subpackage
