--- test/e2e/framework/skip.go
@@ -1,185 +1,0 @@
-/*
-Copyright 2014 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// This file will be removed after switching to use subpackage skipper.
-
-package framework
-
-import (
-	"fmt"
-
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	utilversion "k8s.io/apimachinery/pkg/util/version"
-	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	"k8s.io/client-go/discovery"
-	"k8s.io/client-go/dynamic"
-	clientset "k8s.io/client-go/kubernetes"
-	"k8s.io/kubernetes/pkg/features"
-
-	// TODO: Remove the following imports (ref: https://github.com/kubernetes/kubernetes/issues/81245)
-	"k8s.io/kubernetes/test/e2e/framework/ginkgowrapper"
-	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
-)
-
-func skipInternalf(caller int, format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	log("INFO", msg)
-	ginkgowrapper.Skip(msg, caller+1)
-}
-
-// Skipf skips with information about why the test is being skipped.
-func Skipf(format string, args ...interface{}) {
-	skipInternalf(1, format, args...)
-}
-
-// SkipUnlessAtLeast skips if the value is less than the minValue.
-func SkipUnlessAtLeast(value int, minValue int, message string) {
-	if value < minValue {
-		skipInternalf(1, message)
-	}
-}
-
-// SkipUnlessLocalEphemeralStorageEnabled skips if the LocalStorageCapacityIsolation is not enabled.
-func SkipUnlessLocalEphemeralStorageEnabled() {
-	if !utilfeature.DefaultFeatureGate.Enabled(features.LocalStorageCapacityIsolation) {
-		skipInternalf(1, "Only supported when %v feature is enabled", features.LocalStorageCapacityIsolation)
-	}
-}
-
-// SkipIfMissingResource skips if the gvr resource is missing.
-func SkipIfMissingResource(dynamicClient dynamic.Interface, gvr schema.GroupVersionResource, namespace string) {
-	resourceClient := dynamicClient.Resource(gvr).Namespace(namespace)
-	_, err := resourceClient.List(metav1.ListOptions{})
-	if err != nil {
-		// not all resources support list, so we ignore those
-		if apierrors.IsMethodNotSupported(err) || apierrors.IsNotFound(err) || apierrors.IsForbidden(err) {
-			skipInternalf(1, "Could not find %s resource, skipping test: %#v", gvr, err)
-		}
-		Failf("Unexpected error getting %v: %v", gvr, err)
-	}
-}
-
-// SkipUnlessNodeCountIsAtLeast skips if the number of nodes is less than the minNodeCount.
-func SkipUnlessNodeCountIsAtLeast(minNodeCount int) {
-	if TestContext.CloudConfig.NumNodes < minNodeCount {
-		skipInternalf(1, "Requires at least %d nodes (not %d)", minNodeCount, TestContext.CloudConfig.NumNodes)
-	}
-}
-
-// SkipUnlessNodeCountIsAtMost skips if the number of nodes is greater than the maxNodeCount.
-func SkipUnlessNodeCountIsAtMost(maxNodeCount int) {
-	if TestContext.CloudConfig.NumNodes > maxNodeCount {
-		skipInternalf(1, "Requires at most %d nodes (not %d)", maxNodeCount, TestContext.CloudConfig.NumNodes)
-	}
-}
-
-// SkipIfProviderIs skips if the provider is included in the unsupportedProviders.
-func SkipIfProviderIs(unsupportedProviders ...string) {
-	if ProviderIs(unsupportedProviders...) {
-		skipInternalf(1, "Not supported for providers %v (found %s)", unsupportedProviders, TestContext.Provider)
-	}
-}
-
-// SkipUnlessProviderIs skips if the provider is not included in the supportedProviders.
-func SkipUnlessProviderIs(supportedProviders ...string) {
-	if !ProviderIs(supportedProviders...) {
-		skipInternalf(1, "Only supported for providers %v (not %s)", supportedProviders, TestContext.Provider)
-	}
-}
-
-// SkipUnlessMultizone skips if the cluster does not have multizone.
-func SkipUnlessMultizone(c clientset.Interface) {
-	zones, err := GetClusterZones(c)
-	if err != nil {
-		skipInternalf(1, "Error listing cluster zones")
-	}
-	if zones.Len() <= 1 {
-		skipInternalf(1, "Requires more than one zone")
-	}
-}
-
-// SkipIfMultizone skips if the cluster has multizone.
-func SkipIfMultizone(c clientset.Interface) {
-	zones, err := GetClusterZones(c)
-	if err != nil {
-		skipInternalf(1, "Error listing cluster zones")
-	}
-	if zones.Len() > 1 {
-		skipInternalf(1, "Requires at most one zone")
-	}
-}
-
-// SkipUnlessMasterOSDistroIs skips if the master OS distro is not included in the supportedMasterOsDistros.
-func SkipUnlessMasterOSDistroIs(supportedMasterOsDistros ...string) {
-	if !MasterOSDistroIs(supportedMasterOsDistros...) {
-		skipInternalf(1, "Only supported for master OS distro %v (not %s)", supportedMasterOsDistros, TestContext.MasterOSDistro)
-	}
-}
-
-// SkipUnlessNodeOSDistroIs skips if the node OS distro is not included in the supportedNodeOsDistros.
-func SkipUnlessNodeOSDistroIs(supportedNodeOsDistros ...string) {
-	if !NodeOSDistroIs(supportedNodeOsDistros...) {
-		skipInternalf(1, "Only supported for node OS distro %v (not %s)", supportedNodeOsDistros, TestContext.NodeOSDistro)
-	}
-}
-
-// SkipIfNodeOSDistroIs skips if the node OS distro is included in the unsupportedNodeOsDistros.
-func SkipIfNodeOSDistroIs(unsupportedNodeOsDistros ...string) {
-	if NodeOSDistroIs(unsupportedNodeOsDistros...) {
-		skipInternalf(1, "Not supported for node OS distro %v (is %s)", unsupportedNodeOsDistros, TestContext.NodeOSDistro)
-	}
-}
-
-// SkipUnlessServerVersionGTE skips if the server version is less than v.
-func SkipUnlessServerVersionGTE(v *utilversion.Version, c discovery.ServerVersionInterface) {
-	gte, err := serverVersionGTE(v, c)
-	if err != nil {
-		Failf("Failed to get server version: %v", err)
-	}
-	if !gte {
-		skipInternalf(1, "Not supported for server versions before %q", v)
-	}
-}
-
-// SkipUnlessSSHKeyPresent skips if no SSH key is found.
-func SkipUnlessSSHKeyPresent() {
-	if _, err := e2essh.GetSigner(TestContext.Provider); err != nil {
-		skipInternalf(1, "No SSH Key for provider %s: '%v'", TestContext.Provider, err)
-	}
-}
-
-// serverVersionGTE returns true if v is greater than or equal to the server version.
-func serverVersionGTE(v *utilversion.Version, c discovery.ServerVersionInterface) (bool, error) {
-	serverVersion, err := c.ServerVersion()
-	if err != nil {
-		return false, fmt.Errorf("Unable to get server version: %v", err)
-	}
-	sv, err := utilversion.ParseSemantic(serverVersion.GitVersion)
-	if err != nil {
-		return false, fmt.Errorf("Unable to parse server version %q: %v", serverVersion.GitVersion, err)
-	}
-	return sv.AtLeast(v), nil
-}
-
-// AppArmorDistros are distros with AppArmor support
-var AppArmorDistros = []string{"gci", "ubuntu"}
-
-// SkipIfAppArmorNotSupported skips if the AppArmor is not supported by the node OS distro.
-func SkipIfAppArmorNotSupported() {
-	SkipUnlessNodeOSDistroIs(AppArmorDistros...)
-}
