+++ test/e2e/framework/skipper/skipper.go
@@ -17,57 +17,189 @@ limitations under the License.
 package skipper
 
 import (
-	// TODO: Move function logic from framework after all callers switch to use this package
+	"fmt"
+
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	utilversion "k8s.io/apimachinery/pkg/util/version"
+	utilfeature "k8s.io/apiserver/pkg/util/feature"
+	"k8s.io/client-go/discovery"
+	"k8s.io/client-go/dynamic"
+	clientset "k8s.io/client-go/kubernetes"
+	"k8s.io/kubernetes/pkg/features"
 	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/kubernetes/test/e2e/framework/ginkgowrapper"
+	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
 )
 
+// TestContext should be used by all tests to access common context data.
+var TestContext framework.TestContextType
+
+func skipInternalf(caller int, format string, args ...interface{}) {
+	msg := fmt.Sprintf(format, args...)
+	framework.Logf("INFO", msg)
+	ginkgowrapper.Skip(msg, caller+1)
+}
+
 // Skipf skips with information about why the test is being skipped.
-var Skipf = framework.Skipf
+func Skipf(format string, args ...interface{}) {
+	skipInternalf(1, format, args...)
+}
 
 // SkipUnlessAtLeast skips if the value is less than the minValue.
-var SkipUnlessAtLeast = framework.SkipUnlessAtLeast
+func SkipUnlessAtLeast(value int, minValue int, message string) {
+	if value < minValue {
+		skipInternalf(1, message)
+	}
+}
 
 // SkipUnlessLocalEphemeralStorageEnabled skips if the LocalStorageCapacityIsolation is not enabled.
-var SkipUnlessLocalEphemeralStorageEnabled = framework.SkipUnlessLocalEphemeralStorageEnabled
+func SkipUnlessLocalEphemeralStorageEnabled() {
+	if !utilfeature.DefaultFeatureGate.Enabled(features.LocalStorageCapacityIsolation) {
+		skipInternalf(1, "Only supported when %v feature is enabled", features.LocalStorageCapacityIsolation)
+	}
+}
 
 // SkipIfMissingResource skips if the gvr resource is missing.
-var SkipIfMissingResource = framework.SkipIfMissingResource
+func SkipIfMissingResource(dynamicClient dynamic.Interface, gvr schema.GroupVersionResource, namespace string) {
+	resourceClient := dynamicClient.Resource(gvr).Namespace(namespace)
+	_, err := resourceClient.List(metav1.ListOptions{})
+	if err != nil {
+		// not all resources support list, so we ignore those
+		if apierrors.IsMethodNotSupported(err) || apierrors.IsNotFound(err) || apierrors.IsForbidden(err) {
+			skipInternalf(1, "Could not find %s resource, skipping test: %#v", gvr, err)
+		}
+		framework.Failf("Unexpected error getting %v: %v", gvr, err)
+	}
+}
 
 // SkipUnlessNodeCountIsAtLeast skips if the number of nodes is less than the minNodeCount.
-var SkipUnlessNodeCountIsAtLeast = framework.SkipUnlessNodeCountIsAtLeast
+func SkipUnlessNodeCountIsAtLeast(minNodeCount int) {
+	if TestContext.CloudConfig.NumNodes < minNodeCount {
+		skipInternalf(1, "Requires at least %d nodes (not %d)", minNodeCount, TestContext.CloudConfig.NumNodes)
+	}
+}
 
 // SkipUnlessNodeCountIsAtMost skips if the number of nodes is greater than the maxNodeCount.
-var SkipUnlessNodeCountIsAtMost = framework.SkipUnlessNodeCountIsAtMost
+func SkipUnlessNodeCountIsAtMost(maxNodeCount int) {
+	if TestContext.CloudConfig.NumNodes > maxNodeCount {
+		skipInternalf(1, "Requires at most %d nodes (not %d)", maxNodeCount, TestContext.CloudConfig.NumNodes)
+	}
+}
 
 // SkipIfProviderIs skips if the provider is included in the unsupportedProviders.
-var SkipIfProviderIs = framework.SkipIfProviderIs
+func SkipIfProviderIs(unsupportedProviders ...string) {
+	if framework.ProviderIs(unsupportedProviders...) {
+		skipInternalf(1, "Not supported for providers %v (found %s)", unsupportedProviders, TestContext.Provider)
+	}
+}
 
 // SkipUnlessProviderIs skips if the provider is not included in the supportedProviders.
-var SkipUnlessProviderIs = framework.SkipUnlessProviderIs
+func SkipUnlessProviderIs(supportedProviders ...string) {
+	if !framework.ProviderIs(supportedProviders...) {
+		skipInternalf(1, "Only supported for providers %v (not %s)", supportedProviders, TestContext.Provider)
+	}
+}
 
 // SkipUnlessMultizone skips if the cluster does not have multizone.
-var SkipUnlessMultizone = framework.SkipUnlessMultizone
+func SkipUnlessMultizone(c clientset.Interface) {
+	zones, err := framework.GetClusterZones(c)
+	if err != nil {
+		skipInternalf(1, "Error listing cluster zones")
+	}
+	if zones.Len() <= 1 {
+		skipInternalf(1, "Requires more than one zone")
+	}
+}
 
 // SkipIfMultizone skips if the cluster has multizone.
-var SkipIfMultizone = framework.SkipIfMultizone
+func SkipIfMultizone(c clientset.Interface) {
+	zones, err := framework.GetClusterZones(c)
+	if err != nil {
+		skipInternalf(1, "Error listing cluster zones")
+	}
+	if zones.Len() > 1 {
+		skipInternalf(1, "Requires at most one zone")
+	}
+}
 
 // SkipUnlessMasterOSDistroIs skips if the master OS distro is not included in the supportedMasterOsDistros.
-var SkipUnlessMasterOSDistroIs = framework.SkipUnlessMasterOSDistroIs
+func SkipUnlessMasterOSDistroIs(supportedMasterOsDistros ...string) {
+	if !framework.MasterOSDistroIs(supportedMasterOsDistros...) {
+		skipInternalf(1, "Only supported for master OS distro %v (not %s)", supportedMasterOsDistros, TestContext.MasterOSDistro)
+	}
+}
 
 // SkipUnlessNodeOSDistroIs skips if the node OS distro is not included in the supportedNodeOsDistros.
-var SkipUnlessNodeOSDistroIs = framework.SkipUnlessNodeOSDistroIs
+func SkipUnlessNodeOSDistroIs(supportedNodeOsDistros ...string) {
+	if !framework.NodeOSDistroIs(supportedNodeOsDistros...) {
+		skipInternalf(1, "Only supported for node OS distro %v (not %s)", supportedNodeOsDistros, TestContext.NodeOSDistro)
+	}
+}
 
 // SkipIfNodeOSDistroIs skips if the node OS distro is included in the unsupportedNodeOsDistros.
-var SkipIfNodeOSDistroIs = framework.SkipIfNodeOSDistroIs
+func SkipIfNodeOSDistroIs(unsupportedNodeOsDistros ...string) {
+	if framework.NodeOSDistroIs(unsupportedNodeOsDistros...) {
+		skipInternalf(1, "Not supported for node OS distro %v (is %s)", unsupportedNodeOsDistros, TestContext.NodeOSDistro)
+	}
+}
 
 // SkipUnlessServerVersionGTE skips if the server version is less than v.
-var SkipUnlessServerVersionGTE = framework.SkipUnlessServerVersionGTE
+func SkipUnlessServerVersionGTE(v *utilversion.Version, c discovery.ServerVersionInterface) {
+	gte, err := serverVersionGTE(v, c)
+	if err != nil {
+		framework.Failf("Failed to get server version: %v", err)
+	}
+	if !gte {
+		skipInternalf(1, "Not supported for server versions before %q", v)
+	}
+}
 
 // SkipUnlessSSHKeyPresent skips if no SSH key is found.
-var SkipUnlessSSHKeyPresent = framework.SkipUnlessSSHKeyPresent
+func SkipUnlessSSHKeyPresent() {
+	if _, err := e2essh.GetSigner(TestContext.Provider); err != nil {
+		skipInternalf(1, "No SSH Key for provider %s: '%v'", TestContext.Provider, err)
+	}
+}
+
+// serverVersionGTE returns true if v is greater than or equal to the server version.
+func serverVersionGTE(v *utilversion.Version, c discovery.ServerVersionInterface) (bool, error) {
+	serverVersion, err := c.ServerVersion()
+	if err != nil {
+		return false, fmt.Errorf("Unable to get server version: %v", err)
+	}
+	sv, err := utilversion.ParseSemantic(serverVersion.GitVersion)
+	if err != nil {
+		return false, fmt.Errorf("Unable to parse server version %q: %v", serverVersion.GitVersion, err)
+	}
+	return sv.AtLeast(v), nil
+}
 
 // AppArmorDistros are distros with AppArmor support
-var AppArmorDistros = framework.AppArmorDistros
+var AppArmorDistros = []string{"gci", "ubuntu"}
 
 // SkipIfAppArmorNotSupported skips if the AppArmor is not supported by the node OS distro.
-var SkipIfAppArmorNotSupported = framework.SkipIfAppArmorNotSupported
+func SkipIfAppArmorNotSupported() {
+	SkipUnlessNodeOSDistroIs(AppArmorDistros...)
+}
+
+// RunIfContainerRuntimeIs runs if the container runtime is included in the runtimes.
+func RunIfContainerRuntimeIs(runtimes ...string) {
+	for _, containerRuntime := range runtimes {
+		if containerRuntime == TestContext.ContainerRuntime {
+			return
+		}
+	}
+	skipInternalf(1, "Skipped because container runtime %q is not in %s", TestContext.ContainerRuntime, runtimes)
+}
+
+// RunIfSystemSpecNameIs runs if the system spec name is included in the names.
+func RunIfSystemSpecNameIs(names ...string) {
+	for _, name := range names {
+		if name == TestContext.SystemSpecName {
+			return
+		}
+	}
+	skipInternalf(1, "Skipped because system spec name %q is not in %v", TestContext.SystemSpecName, names)
+}
