+++ test/e2e/storage/testsuites/testdriver.go
@@ -28,7 +28,7 @@ import (
 
 // TestDriver represents an interface for a driver to be tested in TestSuite.
 // Except for GetDriverInfo, all methods will be called at test runtime and thus
-// can use framework.Skipf, framework.Fatal, Gomega assertions, etc.
+// can use e2eskipper.Skipf, framework.Fatal, Gomega assertions, etc.
 type TestDriver interface {
 	// GetDriverInfo returns DriverInfo for the TestDriver. This must be static
 	// information.
