+++ test/e2e_node/container_log_rotation_test.go
@@ -27,6 +27,7 @@ import (
 	kubelogs "k8s.io/kubernetes/pkg/kubelet/logs"
 	kubetypes "k8s.io/kubernetes/pkg/kubelet/types"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2eskipper "k8s.io/kubernetes/test/e2e/framework/skipper"
 
 	"github.com/onsi/ginkgo"
 	"github.com/onsi/gomega"
@@ -45,7 +46,7 @@ var _ = framework.KubeDescribe("ContainerLogRotation [Slow] [Serial] [Disruptive
 	ginkgo.Context("when a container generates a lot of log", func() {
 		ginkgo.BeforeEach(func() {
 			if framework.TestContext.ContainerRuntime != kubetypes.RemoteContainerRuntime {
-				framework.Skipf("Skipping ContainerLogRotation test since the container runtime is not remote")
+				e2eskipper.Skipf("Skipping ContainerLogRotation test since the container runtime is not remote")
 			}
 		})
