+++ test/e2e_node/docker_test.go
@@ -24,6 +24,7 @@ import (
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2eskipper "k8s.io/kubernetes/test/e2e/framework/skipper"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 
 	"github.com/onsi/ginkgo"
@@ -34,7 +35,7 @@ var _ = framework.KubeDescribe("Docker features [Feature:Docker][Legacy:Docker]"
 	f := framework.NewDefaultFramework("docker-feature-test")
 
 	ginkgo.BeforeEach(func() {
-		framework.RunIfContainerRuntimeIs("docker")
+		e2eskipper.RunIfContainerRuntimeIs("docker")
 	})
 
 	ginkgo.Context("when live-restore is enabled [Serial] [Slow] [Disruptive]", func() {
@@ -47,12 +48,12 @@ var _ = framework.KubeDescribe("Docker features [Feature:Docker][Legacy:Docker]"
 			isSupported, err := isDockerLiveRestoreSupported()
 			framework.ExpectNoError(err)
 			if !isSupported {
-				framework.Skipf("Docker live-restore is not supported.")
+				e2eskipper.Skipf("Docker live-restore is not supported.")
 			}
 			isEnabled, err := isDockerLiveRestoreEnabled()
 			framework.ExpectNoError(err)
 			if !isEnabled {
-				framework.Skipf("Docker live-restore is not enabled.")
+				e2eskipper.Skipf("Docker live-restore is not enabled.")
 			}
 
 			ginkgo.By("Create the test pod.")
